# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Suivi VtES API",
    "description": "API du suivi de pratique VtES (collection EN/FR, decks, parties, tournois). Pilote d'une PWA offline-first destinée à être portée sur barrins-project.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/bundles/{bundle_id}/stock": {
      "post": {
        "tags": [
          "catalogue"
        ],
        "summary": "Verser un produit dans la collection",
        "description": "Ajoute le contenu du produit au stock, dans la langue indiquée. Les quantités s'additionnent à l'existant. Non idempotent : deux appels versent deux produits.",
        "operationId": "depositBundle",
        "parameters": [
          {
            "name": "bundle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Bundle Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BundleDeposit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CardCopyRead"
                  },
                  "title": "Response Depositbundle"
                }
              }
            }
          },
          "404": {
            "description": "Ressource introuvable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflit avec l'état des données (doublon, stock, règle).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/stock": {
      "get": {
        "tags": [
          "stock"
        ],
        "summary": "Liste la collection",
        "operationId": "listStock",
        "parameters": [
          {
            "name": "language_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Language Code"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CardCategory"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 80
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CardCopyRead"
                  },
                  "title": "Response Liststock"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "stock"
        ],
        "summary": "Déclare une carte dans une langue",
        "operationId": "createStockEntry",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardCopyCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardCopyRead"
                }
              }
            }
          },
          "404": {
            "description": "Ressource introuvable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflit avec l'état des données (doublon, stock, règle).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/stock/{card_id}/{language_code}": {
      "get": {
        "tags": [
          "stock"
        ],
        "summary": "Lit une entrée de collection",
        "operationId": "getStockEntry",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Card Id"
            }
          },
          {
            "name": "language_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Language Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardCopyRead"
                }
              }
            }
          },
          "404": {
            "description": "Ressource introuvable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "stock"
        ],
        "summary": "Modifie une entrée de collection",
        "operationId": "updateStockEntry",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Card Id"
            }
          },
          {
            "name": "language_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Language Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardCopyUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardCopyRead"
                }
              }
            }
          },
          "404": {
            "description": "Ressource introuvable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflit avec l'état des données (doublon, stock, règle).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "stock"
        ],
        "summary": "Retire une entrée de collection",
        "description": "Refusé (409) tant que des decks utilisent l'entrée.",
        "operationId": "deleteStockEntry",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Card Id"
            }
          },
          {
            "name": "language_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Language Code"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Ressource introuvable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflit avec l'état des données (doublon, stock, règle).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BundleDeposit": {
        "properties": {
          "language_code": {
            "type": "string",
            "maxLength": 8,
            "minLength": 1,
            "title": "Language Code"
          },
          "count": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Count",
            "description": "Nombre de produits identiques versés d'un coup.",
            "default": 1
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "language_code"
        ],
        "title": "BundleDeposit",
        "description": "Versement du contenu d'un produit dans la collection.\n\nLe produit fixe les cartes et leurs exemplaires ; il ne manque que la\nlangue de ce qui a été acheté, et le nombre de produits identiques."
      },
...
      "CardCategory": {
        "type": "string",
        "enum": [
          "crypt",
          "library"
        ],
        "title": "CardCategory",
        "description": "Les deux moitiés d'un deck VtES (CLAUDE.md §5)."
      },
      "CardCopyCreate": {
        "properties": {
          "card_id": {
            "type": "integer",
            "title": "Card Id"
          },
          "language_code": {
            "type": "string",
            "maxLength": 8,
            "minLength": 1,
            "title": "Language Code"
          },
          "quantity_owned": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Quantity Owned",
            "default": 0
          },
          "proxy_allowed": {
            "type": "boolean",
            "title": "Proxy Allowed",
            "default": false
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "card_id",
          "language_code"
        ],
        "title": "CardCopyCreate",
        "description": "Déclaration d'une entrée de collection."
      },
      "CardCopyRead": {
        "properties": {
          "card_id": {
            "type": "integer",
            "title": "Card Id"
          },
          "language_code": {
            "type": "string",
            "title": "Language Code"
          },
          "quantity_owned": {
            "type": "integer",
            "title": "Quantity Owned",
            "description": "Nombre d'exemplaires réellement possédés."
          },
          "proxy_allowed": {
            "type": "boolean",
            "title": "Proxy Allowed",
            "description": "Autorise à jouer cette carte en proxy dans cette langue, sans la posséder. Une entrée avec 0 exemplaire possédé et le proxy autorisé est le cas normal d'une carte jouée en proxy."
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "card": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CardSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "card_id",
          "language_code",
          "quantity_owned",
          "proxy_allowed"
        ],
        "title": "CardCopyRead",
        "description": "Exemplaires possédés d'une carte dans une langue."
      },
      "CardCopyUpdate": {
        "properties": {
          "quantity_owned": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Quantity Owned"
          },
          "proxy_allowed": {
            "type": "boolean",
            "title": "Proxy Allowed"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CardCopyUpdate",
        "description": "Modification d'une entrée de collection.\n\nLa carte et la langue forment la clé : elles ne se modifient pas, on crée\nune autre entrée. `quantity_owned` et `proxy_allowed` sont facultatifs mais\nnon nullables (colonnes NOT NULL) ; `notes`, lui, accepte `null` pour\neffacer la note."
      },
...
      "CardSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "vekn_id": {
            "type": "integer",
            "title": "Vekn Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "$ref": "#/components/schemas/CardCategory"
          },
          "clan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClanRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "capacity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity"
          },
          "group_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Code"
          },
          "advanced": {
            "type": "boolean",
            "title": "Advanced",
            "default": false
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url",
            "description": "Scan de la carte (version anglaise de référence)."
          }
        },
        "type": "object",
        "required": [
          "id",
          "vekn_id",
          "name",
          "category"
        ],
        "title": "CardSummary",
        "description": "Vue courte d'une carte, pour les listes et l'autocomplétion."
      },
...
      "ClanRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "abbrev": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abbrev"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "ClanRead",
        "description": "Clan."
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Explication lisible de l'échec."
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "Corps d'une réponse d'erreur métier."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}